+++ lib/mongoid/criteria.rb
@@ -1,6 +1,7 @@
 # encoding: utf-8
 require "mongoid/criterion/inspection"
 require "mongoid/criterion/findable"
+require "mongoid/criterion/marshalable"
 require "mongoid/criterion/modifiable"
 require "mongoid/criterion/scoping"
 
@@ -18,6 +19,7 @@ module Mongoid
     include Origin::Queryable
     include Criterion::Inspection
     include Criterion::Findable
+    include Criterion::Marshalable
     include Criterion::Modifiable
     include Criterion::Scoping