+++ spec/mongoid/criteria_spec.rb
@@ -2234,7 +2234,7 @@ describe Mongoid::Criteria do
     end
   end
 
-  describe "$gt" do
+  describe "#gt" do
 
     let!(:match) do
       Band.create(member_count: 5)
@@ -2253,7 +2253,7 @@ describe Mongoid::Criteria do
     end
   end
 
-  describe "$gte" do
+  describe "#gte" do
 
     let!(:match) do
       Band.create(member_count: 5)