+++ CHANGELOG.md
@@ -145,6 +145,9 @@ For instructions on upgrading to newer versions, visit
 
 * \#2630 Fix cascading when the metadata exists but no cascade defined.
 
+* \#2625 Fix `Marshal.dump` and `Marshal.load` of proxies and criteria
+  objects.
+
 * \#2619 Fixed the classes returned by `observed_classes` on an observer
   when it is observing custom models.